billing_admin/requirements.txt
django==4.1.7
python-dotenv==1.0.0
django-split-settings==1.2.0
psycopg2-binary==2.9.5
gunicorn==20.1.0